requirements.txt
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
evdev==1.9.0
h11==0.14.0
idna==3.10
imageio==2.37.0
lazy_loader==0.4
MouseInfo==0.1.3
networkx==3.4.2
numpy==2.2.3
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
PyAutoGUI==0.9.54
pygame==2.6.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tifffile==2025.2.18
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
